+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+### Added
+- Add optional custom_prompt arg to the Text2CypherRetriever class.
+  
 ## 0.3.1
 
 ### Fixed