+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+## 0.4.0
+
 ### Added
 - Add optional custom_prompt arg to the Text2CypherRetriever class.